:bug: fix: -Handled conflicting files in the archive folder by deleting the existing file. -Checked whether the file to be archived is still available

+++ tofhir-engine/src/main/scala/io/tofhir/engine/execution/FileStreamInputArchiver.scala
@@ -184,8 +184,20 @@ object FileStreamInputArchiver {
 
     // create parent directories if not exists
     archiveFile.getParentFile.mkdirs()
-    // move file to archive folder
-    Files.move(file.toPath, archiveFile.toPath)
+
+    // Check if the parent folder already contains a file with the same name. Delete, if yes
+    val parentDirectory: File = archiveFile.getParentFile
+    parentDirectory.listFiles().find(f => f.getName.contentEquals(archiveFile.getName)) match {
+      case None =>
+      case Some(file) => file.delete()
+    }
+
+    // We need to check whether the input file still exists.
+    // It might not exist in the following scenario: It has already been archived, the system is restarted and archiving starts from offset 0 and
+    // tries to rearchive the file.
+    if (!archiveFile.exists()) {
+      Files.move(file.toPath, archiveFile.toPath)
+    }
   }
 }
